==> 03_data_prep.ipynb <==
# Preparação dos Dados   

## 1 - Bibliotecas e Dados

import pandas as pd
import numpy as np
import matplotlib.pyplot as  plt
import seaborn as sns
import sklearn
from sklearn.preprocessing import MinMaxScaler
from scipy.stats import ttest_ind, mannwhitneyu
import scipy.stats as stats
import statsmodels.api as sm
import statsmodels.api as sm
import featuretools as ft
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
import pickle

df = pd.read_csv('data/df_EDA.csv')

## 2 - Tratamento de Missing Values

df.isna().sum()

### 2.1 - Released Year

Considerando que não teria como imputar o ano que o filme foi produzido, mas também que só tem um filme sem essa informação, vou inserir essa informação

df[df['Released_Year'].isna()]

Apolo 13 é um filme de 1995, logo vou inserir esse valor

df.loc[df['Released_Year'].isna(), 'Released_Year'] = 1995

### 2.2 - Released Month e Released Day e Released Date

Por conta do alto valor de missings nessas variáveis, vou retiirá-las do modelo, assim evita de perder muita informação

df = df.drop(columns=['Released_Month', 'Released_Day', 'release_date'])

### 2.3 - Certificate

Para manter uma coerência nas classificações, vou utilizar a moda por gênero, assim gêneros com classificações mais severas como Terror, segurão essas linhas, e mais brandos como romances e familiares também terão uma lógica

df['Main_Genre'] = df['Genre'].str.split(',').str[0]
df['Certificate'] = df.groupby('Main_Genre', group_keys=False)['Certificate'] \
                      .apply(lambda x: x.fillna(x.mode()[0] if not x.mode().empty else 'Not Rated'))

### 2.4 - Meta_score

Existem duas opções claras: Imputar diretamente pela mediana, ou usar IMDB_rating como base, vou verificar a correlação (já vista no mapa de correlação na EDA)

corr = df.loc[df['Meta_score'].notna(), 'Meta_score'].corr(
    df.loc[df['Meta_score'].notna(), 'IMDB_Rating']
)
print(f"Correlação entre IMDB_Rating e Meta_score: {corr:.2f}")

Como a correlação é baixa, vou usar a mediana

df['Meta_score'] = df['Meta_score'].fillna(df['Meta_score'].median())

### 2.5 Gross, Budget e ROI (e seus logs)

corr_bg = df.loc[df['budget'].notna() & df['Gross'].notna(), 'budget'] \
           .corr(df.loc[df['budget'].notna() & df['Gross'].notna(), 'Gross'])

print(f"Correlação entre budget e Gross: {corr_bg:.2f}")

Levando em consideração a alta correlação, vou utilizá-los para realizar o imput onde for possível

mask_budget = df['budget'].isna() & df['Gross'].notna()
if mask_budget.any():
    model_budget = LinearRegression()
    train_data = df[df['budget'].notna() & df['Gross'].notna()]
    model_budget.fit(train_data[['Gross']], train_data['budget'])
    df.loc[mask_budget, 'budget'] = model_budget.predict(df.loc[mask_budget, ['Gross']])

mask_gross = df['Gross'].isna() & df['budget'].notna()
if mask_gross.any():
    model_gross = LinearRegression()
    train_data = df[df['Gross'].notna() & df['budget'].notna()]
    model_gross.fit(train_data[['budget']], train_data['Gross'])
    df.loc[mask_gross, 'Gross'] = model_gross.predict(df.loc[mask_gross, ['budget']])
    df.loc[mask_gross, 'Gross'] = model_gross.predict(df.loc[mask_gross, ['budget']])

df.isna().sum()

dos 169 dados ausentes em Gross e 112 em Budget, sobraram apenas 36 em cada. Vou preencher o restante com a mediana e em seguida refazer as colunas de ROI e as transformações logarítimicas

df['budget'] = df['budget'].fillna(df['budget'].median())
df['Gross'] = df['Gross'].fillna(df['Gross'].median())

df['ROI'] = (df['Gross'] - df['budget']) / df['budget']

df['Gross_log'] = np.log1p(df['Gross'])
df['budget_log'] = np.log1p(df['budget'])
df['ROI_log'] = np.log1p(df['ROI'])

df.isna().sum()

df_split = df.to_csv("data/df_split.csv", index=False)

## 3 - Separação em Treino e Teste (Split)

Considerando que a proposta é prever a nota do imdb, vou isolar ela e fazer o split na proporção 70-30 por ser um dataset pequeno.
O modelo será de classificação, não vou utilizar uma estratificação pois algumas notas são muito raras, não valendo a pena criar uma categoria de 'outros' apenas para elas, já que podem ser muito altas ou muito baixas

df_split['IMDB_Rating'] = df_split['IMDB_Rating'].astype('category')

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42
)

## 4 - Feature Engineering

Para a engenharia de variáveis, vou usar o feature tools, pacote que permite fazer o processo de forma automatizada, assim extraindo mais informações do dataset

es = ft.EntitySet(id='movies')

es = es.add_dataframe(
    dataframe_name='movies_train',
    dataframe=X_train,
    index='index',  
)

feature_matrix, feature_defs = ft.dfs(
    entityset=es,
    target_dataframe_name='movies_train',
    max_depth=2,   
    verbose=1
)

feature_matrix.head()

O feature tools me deu mais duas variáveis, número de palavras e caracteres na sinópse, agora vou aplicar a mesma coisa no teste

X_test['NUM_CHARACTERS(Overview)'] = X_test['Overview'].apply(len)
X_test['NUM_WORDS(Overview)'] = X_test['Overview'].apply(lambda x: len(x.split()))

with open('X_train.pkl', 'wb') as f:
    pickle.dump(X_train, f)
with open('y_train.pkl', 'wb') as f:
    pickle.dump(y_train, f)
with open('X_test.pkl', 'wb') as f:
    pickle.dump(X_test, f)
with open('y_test.pkl', 'wb') as f:
    pickle.dump(y_test, f)
