==> 04_modelo.ipynb <==
# Modelagem

# 1 - Bibliotecas e Dados 

import pickle
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_squared_error, r2_score
import joblib


with open('data/X_train.pkl', 'rb') as f:
    X_train = pickle.load(f)
with open('data/y_train.pkl', 'rb') as f:
    y_train = pickle.load(f)
with open('data/X_test.pkl', 'rb') as f:
    X_test = pickle.load(f)
with open('data/y_test.pkl', 'rb') as f:
    y_test = pickle.load(f)

## 2 - Modelos

Vou utilizarum Random Forest, a ideia é que ele é um modelo que assume não linearidade, lida bem com bases pequenas evitando um overfitting e vai poder trabalhar com os tipos de variáveis do nosso modelo
Em conjunto, vou aplicar um Gradient Boosting para ir corrigindo o gradiente do erro
Como métricas de erros vou utilizar o R² que vai mostrar de forma geral o quão bem o modelo consegue capturar as tendências em conjunto do MSE para penalizar erros grandes que podem ser mascarados pelo R²

# Garantir DataFrame
X_train = pd.DataFrame(X_train)
X_test = pd.DataFrame(X_test)

# One-hot encoding
X_train_enc = pd.get_dummies(X_train)
X_test_enc = pd.get_dummies(X_test)

X_test_enc = X_test_enc.reindex(columns=X_train_enc.columns, fill_value=0)

X_train_enc.replace([np.inf, -np.inf], np.nan, inplace=True)
X_test_enc.replace([np.inf, -np.inf], np.nan, inplace=True)

X_train_enc.fillna(0, inplace=True)
X_test_enc.fillna(0, inplace=True)

models = {
    'Random Forest': RandomForestRegressor(n_estimators=100, max_depth=3, min_samples_leaf=300, random_state=7),
    'Gradient Boosting': GradientBoostingRegressor(n_estimators=100, max_depth=2, min_samples_leaf=300, learning_rate=0.05, random_state=7)
} # Limitações para evitar overfiting

for name, model in models.items():
    model.fit(X_train_enc, y_train)
    y_pred = model.predict(X_test_enc)
    mse = mean_squared_error(y_test, y_pred)
    r2 = r2_score(y_test, y_pred)
    print(f"{name}: MSE = {mse:.4f}, R2 = {r2:.4f}")

Com isso temos que o GB é um modelo de classificação adequado, seu erro médio é baixo e o R² elevado, mas nao ao ponto de indicar overfitting, apontam um sucesso do modelo em conjunto com o baixo MSE

gb_model = GradientBoostingRegressor(
    n_estimators=100,
    max_depth=2,
    min_samples_leaf=300,
    learning_rate=0.05,
    random_state=7
)

gb_model.fit(X_train_enc, y_train)

y_pred = gb_model.predict(X_test_enc)
mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Gradient Boosting: MSE = {mse:.4f}, R2 = {r2:.4f}")

# Salvar em .pkl
joblib.dump(gb_model, "modelo_final.pkl")

## 3 - Classificando o filme The Shawshank Redemption

novo_filme = {
    'Series_Title': ['The Shawshank Redemption'],
    'Released_Year': [1994],
    'Certificate': ['A'],
    'Runtime': ['142 min'],
    'Genre': ['Drama'],
    'Overview': ['Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency.'],
    'Meta_score': [80.0],
    'Director': ['Frank Darabont'],
    'Star1': ['Tim Robbins'],
    'Star2': ['Morgan Freeman'],
    'Star3': ['Bob Gunton'],
    'Star4': ['William Sadler'],
    'No_of_Votes': [2343110],
    'Gross': ['28,341,469'] 
}

novo_df = pd.DataFrame(novo_filme)

novo_df['Gross'] = novo_df['Gross'].str.replace(',', '').astype(float)
novo_df['Runtime'] = novo_df['Runtime'].str.replace(' min','').astype(float)

# One-hot encoding
novo_df_enc = pd.get_dummies(novo_df)

novo_df_enc = novo_df_enc.reindex(columns=X_train_enc.columns, fill_value=0)

gbm_pred = models['Gradient Boosting'].predict(novo_df_enc)
print("Gradient Boosting prevê nota:", gbm_pred[0])

Considerando que The Shawshank Redemption tem nota 9.3 no IMDB, o erro foi de aproximadamente 1 ponto
